requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
azure-core==1.27.1
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
cssselect2==0.7.0
Django==4.2.2
django-storages==1.13.2
Django-Verify-Email==2.0.3
html5lib==1.1
idna==3.4
isodate==0.6.1
joblib==1.2.0
lxml==4.9.2
numpy==1.24.3
oscrypto==1.3.0
pandas==2.0.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
threadpoolctl==3.1.0
tinycss2==1.2.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
